Add macro to inject WASM test boilerplate code
This adds a macro to inject some of the boilerplate code that WASM tests require, such as various includes and a function to create and get a webgl2 context

Related to #25
